[rf] refactor checkRegistration to usercontext
